# backend/codegen/src/main/java/eu/raweceek/codegen/api/SessionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package eu.raweceek.codegen.api;

import eu.raweceek.codegen.models.SessionDto;
import eu.raweceek.codegen.models.SessionsCountdownGet200Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0-SNAPSHOT")
@Validated
@Tag(name = "sessions", description = "the sessions API")
public interface SessionsApi {

    /**
     * GET /sessions/countdown
     *
     * @return Countdown until the next session by startTime (status code 200)
     *         or Next session not found (status code 404)
     */
    @Operation(
        operationId = "sessionsCountdownGet",
        responses = {
            @ApiResponse(responseCode = "200", description = "Countdown until the next session by startTime", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SessionsCountdownGet200Response.class))
            }),
            @ApiResponse(responseCode = "404", description = "Next session not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sessions/countdown",
        produces = { "application/json" }
    )
    
    ResponseEntity<SessionsCountdownGet200Response> sessionsCountdownGet(
        
    );


    /**
     * GET /sessions
     *
     * @return List of upcoming sessions (status code 200)
     */
    @Operation(
        operationId = "sessionsGet",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of upcoming sessions", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = SessionDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sessions",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<SessionDto>> sessionsGet(
        
    );


    /**
     * GET /sessions/next
     *
     * @return Next session by startTime (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "sessionsNextGet",
        responses = {
            @ApiResponse(responseCode = "200", description = "Next session by startTime", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SessionDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sessions/next",
        produces = { "application/json" }
    )
    
    ResponseEntity<SessionDto> sessionsNextGet(
        
    );

}