# backend/spec/swagger.yaml
openapi: 3.0.3
info:
  title: Rawe Ceek API
  version: 1.2.0
servers:
  - url: https://raweceek.eu/api
tags:
  - name: Session
  - name: System
paths:
  /sessions:
    get:
      tags:
        - Session
      responses:
        '200':
          description: List of upcoming sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionDto"
  /sessions/next:
    get:
      tags:
        - Session
      responses:
        '200':
          description: Next session by startTime
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/SessionDto"
        '404':
          description: Not found
  /sessions/countdown:
    get:
      tags:
        - Session
      responses:
        '200':
          description: Countdown until the next session by startTime
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: "#/components/schemas/SessionDto"
                  countdowns:
                    type: array
                    items:
                      $ref: "#/components/schemas/CountdownDto"
        '404':
          description: Next session not found
  /system/health:
    get:
      tags:
        - System
      description: Returns the health status of the service
      responses:
        '200':
          description: Service is healthy
        default:
          description: Service is not healthy
components:
  schemas:
    SessionDto:
      type: object
      required:
        - summary
        - location
        - startTime
      properties:
        summary:
          type: string
          example: 🏁 GRAND PRIX NAME 2024 - Race
        location:
          type: string
          example: United Arab Emirates
        startTime:
          type: string
          example: 2024-12-08T13:00Z
    CountdownDto:
      type: object
      required:
        - value
        - unit
      properties:
        value:
          type: double
          example: 3.95
        unit:
          type: string
          example: ceeks